Refactor router and add course materials download functionality
The router in Router.js was modified to include a new path
"/courses/material/:id/:code", allowing for the code of the course
material to be included in the URL. Additionally, in CouresDetails.jsx,
a Link component was updated to include both the id and code parameters
when navigating to the course material page. In CouresMaterial.jsx, the
functionality to download course materials was implemented, utilizing a
list of materials imported from the data.js file. The commit refactors
the router and enables users to download course materials.

Refactor router, enable course materials download

Modify Router.js to include new URL path. Update
Link component in CouresDetails.jsx to navigate to course
material page with id and code parameters. Implement
course material download functionality in CouresMaterial.jsx.
Addresses #issueNumber